pyproject.toml
[tool.poetry]
name = "expo"
version = "0.1.0"
description = "Explanation of ODCM Initiative"
authors = ["Elena Romanenko <eromanenko@unibz.it>"]
readme = "README.md"
packages = [
    {include = "expose"}
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.21.1"
requests = "^2.28.2"
fastapi = "^0.95.1"
python-decouple = "^3.8"
python-multipart = "^0.0.6"
pygithub = "^1.58.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
